Backend/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.2
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
mysqlclient==2.2.7
SQLAlchemy==2.0.43
typing_extensions==4.14.1
Werkzeug==3.1.3
